# docs/openapi.json
{"openapi": "3.1.0", "info": {"title": "Nifty 100 API", "description": "REST API for Nifty 100 Financial Analytics Platform", "version": "1.0.0"}, "paths": {"/api/v1/health": {"get": {"summary": "Health Check", "operationId": "health_check_api_v1_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/companies/": {"get": {"tags": ["Companies"], "summary": "Get Companies", "operationId": "get_companies_api_v1_companies__get", "parameters": [{"name": "sector", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Sector"}}, {"name": "search", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Search"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/companies/{ticker}": {"get": {"tags": ["Companies"], "summary": "Get Company Profile", "operationId": "get_company_profile_api_v1_companies__ticker__get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/companies/{ticker}/pl": {"get": {"tags": ["Companies"], "summary": "Get Pl History", "operationId": "get_pl_history_api_v1_companies__ticker__pl_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}, {"name": "from_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "From Year"}}, {"name": "to_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "To Year"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/companies/{ticker}/bs": {"get": {"tags": ["Companies"], "summary": "Get Bs History", "operationId": "get_bs_history_api_v1_companies__ticker__bs_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}, {"name": "from_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "From Year"}}, {"name": "to_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "To Year"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/companies/{ticker}/cashflow": {"get": {"tags": ["Companies"], "summary": "Get Cashflow History", "operationId": "get_cashflow_history_api_v1_companies__ticker__cashflow_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}, {"name": "from_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "From Year"}}, {"name": "to_year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "To Year"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/companies/{ticker}/ratios": {"get": {"tags": ["Companies"], "summary": "Get Ratio History", "operationId": "get_ratio_history_api_v1_companies__ticker__ratios_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}, {"name": "year", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Year"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/companies/{ticker}/tearsheet": {"get": {"tags": ["Companies"], "summary": "Download Tearsheet", "operationId": "download_tearsheet_api_v1_companies__ticker__tearsheet_get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/screener/": {"get": {"tags": ["Screener"], "summary": "Screen Companies", "operationId": "screen_companies_api_v1_screener__get", "parameters": [{"name": "min_roe", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "Minimum Return on Equity (%)", "title": "Min Roe"}, "description": "Minimum Return on Equity (%)"}, {"name": "max_de", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "Maximum Debt to Equity", "title": "Max De"}, "description": "Maximum Debt to Equity"}, {"name": "min_fcf", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "Minimum Free Cash Flow (Cr)", "title": "Min Fcf"}, "description": "Minimum Free Cash Flow (Cr)"}, {"name": "sector", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "description": "Filter by Broad Sector", "title": "Sector"}, "description": "Filter by Broad Sector"}, {"name": "min_rev_cagr_5yr", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "Minimum 5Y Revenue CAGR", "title": "Min Rev Cagr 5Yr"}, "description": "Minimum 5Y Revenue CAGR"}, {"name": "min_pat_cagr_5yr", "in": "query", "required": false, "schema": {"anyOf": [{"type": "number"}, {"type": "null"}], "description": "Minimum 5Y PAT CAGR", "title": "Min Pat Cagr 5Yr"}, "description": "Minimum 5Y PAT CAGR"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/sectors/": {"get": {"tags": ["Sectors"], "summary": "Get Sectors", "operationId": "get_sectors_api_v1_sectors__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/sectors/{sector_name}/companies": {"get": {"tags": ["Sectors"], "summary": "Get Companies In Sector", "operationId": "get_companies_in_sector_api_v1_sectors__sector_name__companies_get", "parameters": [{"name": "sector_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Sector Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/peers/{group_name}": {"get": {"tags": ["Peers"], "summary": "Get Peer Group", "operationId": "get_peer_group_api_v1_peers__group_name__get", "parameters": [{"name": "group_name", "in": "path", "required": true, "schema": {"type": "string", "title": "Group Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/valuation/{ticker}": {"get": {"tags": ["Valuation"], "summary": "Get Valuation", "operationId": "get_valuation_api_v1_valuation__ticker__get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/v1/portfolio/stats": {"get": {"tags": ["Portfolio"], "summary": "Get Portfolio Stats", "operationId": "get_portfolio_stats_api_v1_portfolio_stats_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/v1/documents/{ticker}": {"get": {"tags": ["Documents"], "summary": "Get Documents", "operationId": "get_documents_api_v1_documents__ticker__get", "parameters": [{"name": "ticker", "in": "path", "required": true, "schema": {"type": "string", "title": "Ticker"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}